code/2025-08-02-langchain-structured-fastapi/pyproject.toml
[tool.poetry]
name = "langchain-structured-fastapi"
version = "0.1.0"
description = "Demo API showing LangChain structured outputs with FastAPI"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
langchain = "^0.3.0"
langchain-openai = "^0.2.0"
pydantic = "^2.0.0"
uvicorn = {extras = ["standard"], version = "^0.30.0"}
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
httpx = "^0.27.0"
pytest-asyncio = "^0.24.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
